// src/enums/iconsEnum.ts
export enum IconsEnum {
  AlignCamera = "AlignCamera",
  Award = "Award",
  ArrowUp = "ArrowUp",
  CheckLine = "CheckLine",
  CheckList = "CheckList",
  ChevronDown = "ChevronDown",
  Error = "Error",
  Facebook = "Facebook",
  Instagram = "Instagram",
  Happy = "Happy",
  Logo = "Logo",
  Menu = "Menu",
  MenuBack = "MenuBack",
  MenuClose = "MenuClose",
  Sad = "Sad",
  Surprised = "Surprised",
  Disgusted = "Disgusted",
  Angry = "Angry",
  Prev = "Prev",
  Next = "Next",
  Play = "Play",
  Like = "Like",
  Swipe = "Swipe",
  Eurofarma = "Eurofarma",
}