// src/enums/screenNamesEnum.ts
export enum ScreenNamesEnum {
  Language = "Language",
  Home = "Home",
  TutorialStart = "TutorialStart",
  Social = "Social",
  Drawer = "Drawer",
  Dashboard = "Dashboard",
  Settings = "Settings",
  Splash = "Splash",
  Stats = "Stats",
  Tutorial = "Tutorial",
  TermsAndConditions = "TermsAndConditions",
  PrivacyPolicy = "PrivacyPolicy",
  TutorialCompleted = "TutorialCompleted",
  Help = "Help"
}